Implement review system
This patch adds a new route for submit which allows the match number to
be specified in the URL, and will cause the score data to be loaded and
automatically populated in the form.

By providing a simple wizard with a selection for the match number,
which then takes the user to the new submit route, a basic reviewing
system is in place.